pyproject.toml
[tool.poetry]
name = "Prompt2YOLO"
version = "0.2.1"
description = "A pipeline for automatically generating training data using prompts and fine-tuning YOLO models."
authors = ["Joseph Wang <joseph.wang@instai.co>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/instAI/Prompt2YOLO"

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.4.1"
numpy = "^1.21.0"
opencv-python = "^4.5.3"
Pillow = "^10.4.0"
matplotlib = "^3.4.2"
tqdm = "^4.61.2"
transformers = "^4.12.0"
datasets = "^1.11.0"
seaborn = "^0.13.1"
albumentations = "^1.0.0"
requests = "^2.26.0"
pandas = "^2.1.0"
rclone = "^0.4.4"
diffusers = "^0.30.3"
ultralytics = "^8.2.101"
boto3 = "^1.35.27"
python-dotenv = "^1.0.1"
peft = "^0.7.0"
awscli = "^1.34.29"
huggingface-hub = "^0.25.2"
streamlit = "^1.39.0"
pyyaml = "^6.0.2"
scikit-learn = "^1.5.2"
pytest-cov = "^6.0.0"
ansi2html = "^1.9.2"
filelock = "^3.16.1"
streamlit-autorefresh = "^1.0.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
black = "^21.7b0"
flake8 = "^3.9.2"
pre-commit = "^2.14.0"
ipykernel = "^6.0.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[scripts]
train-yolo = "src.main:train_yolo"
generate-data = "src.main:generate_data"
sync-drive = "src.drive_sync:sync_google_drive"
